pyproject.toml
[project]
name = "livro-dos-sonhos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
]
